% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{file_exists_file}
\alias{file_exists_file}
\title{Check if a user file exists on the local file system}
\usage{
file_exists_file(user_id, file_name, base_dir)
}
\arguments{
\item{user_id}{The user ID}

\item{file_name}{The file name to check}

\item{base_dir}{The base directory for user data}
}
\value{
Logical TRUE/FALSE
}
\description{
Check if a user file exists on the local file system
}
