% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_from_file}
\alias{remove_from_file}
\title{Remove a File from the Local File System}
\usage{
remove_from_file(user_id, file_name, base_dir)
}
\arguments{
\item{user_id}{A string specifying the unique user ID.}

\item{file_name}{The name of the file to remove (e.g., `"account_tree.Rds"`).}

\item{base_dir}{The root directory containing all user folders. The full
path is constructed as `file.path(base_dir, user_id, file_name)`.}
}
\value{
Returns `TRUE` if the file was successfully removed, or `FALSE` if
  the file did not exist or could not be removed. A warning is issued if the
  file is not found.
}
\description{
Removes a specified file from a user's directory on the local file system.
This is a plugin for the `"file"` backend, used via the generic
interface `remove_user_file()`.
}
\details{
If the file does not exist, a warning is issued but no error is thrown.
}
\examples{
\dontrun{
remove_from_file("user123", "account_tree.Rds", base_dir = "user_data")
}

}
\seealso{
[save_to_file()], [load_from_file()], [remove_user_file()]
}
